+++ web-ui/src/components/root-components/AppSidebar.tsx
@@ -97,7 +97,7 @@ export const AppSidebar = ({ isTablet, setMobileOpen }: AppSidebarProps) => {
             © 2026 Adamax-Works
           </Typography>
           <Typography variant="caption" display="block" color="text.secondary">
-            v5.8.4
+            v5.9.0
           </Typography>
         </Box>
       </Box>